remove unnecessary fields from weatherResponse and fix type converter errors

+++ app/src/main/java/com/jobinlawrance/weather/data/network/model/Clouds.kt
@@ -1,5 +1,5 @@
 package com.jobinlawrance.weather.data.network.model
 
 data class Clouds(
-	val all: Int? = null
+	val all: Double? = null
 )
